Refactor import functions to static methods
