+++ libraries/classes/Plugins/Import/README
@@ -12,6 +12,7 @@ gettext mechanism, see https://wiki.phpmyadmin.net/pma/Gettext_for_developers.
  */
 namespace PhpMyAdmin\Plugins\Import;
 
+use PhpMyAdmin\Import;
 use PhpMyAdmin\Plugins\ImportPlugin;
 
 /**
@@ -104,7 +105,7 @@ class Import[Name] extends ImportPlugin
 
         $buffer = '';
         while (! ($finished && $i >= $len) && ! $error && ! $timeout_passed) {
-            $data = PMA_importGetNextChunk();
+            $data = Import::getNextChunk();
             if ($data === false) {
                 // subtract data we didn't handle yet and stop processing
                 $GLOBALS['offset'] -= strlen($buffer);
@@ -116,10 +117,10 @@ class Import[Name] extends ImportPlugin
                 $buffer .= $data;
             }
             // PARSE $buffer here, post sql queries using:
-            PMA_importRunQuery($sql, $verbose_sql_with_comments, $sql_data);
+            Import::runQuery($sql, $verbose_sql_with_comments, $sql_data);
         } // End of import loop
         // Commit any possible data in buffers
-        PMA_importRunQuery('', '', $sql_data);
+        Import::runQuery('', '', $sql_data);
     }
