+++ test/classes/plugin/import/ImportLdiTest.php
@@ -19,7 +19,6 @@ $GLOBALS['plugin_param'] = "table";
  * Include to test.
  */
 require_once 'libraries/database_interface.inc.php';
-require_once 'libraries/import.lib.php';
 require_once 'test/PMATestCase.php';
 
 /**
