+++ test/classes/plugin/import/ImportMediawikiTest.php
@@ -17,7 +17,6 @@ $GLOBALS['server'] = 0;
  * Include to test.
  */
 require_once 'libraries/database_interface.inc.php';
-require_once 'libraries/import.lib.php';
 require_once 'test/PMATestCase.php';
 
 /**
