+++ test/classes/plugin/import/ImportShpTest.php
@@ -15,7 +15,6 @@ use PhpMyAdmin\File;
 $GLOBALS['server'] = 0;
 
 require_once 'libraries/database_interface.inc.php';
-require_once 'libraries/import.lib.php';
 require_once 'test/PMATestCase.php';
 
 /**
