+++ test/classes/plugin/import/ImportXmlTest.php
@@ -18,7 +18,6 @@ $GLOBALS['server'] = 0;
  * Include to test.
  */
 require_once 'libraries/database_interface.inc.php';
-require_once 'libraries/import.lib.php';
 require_once 'test/PMATestCase.php';
 
 /**
